+++ ricaun.Revit.Installation.Tests/RevitUtils_Tests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using System;
+using System.IO;
 using System.Reflection;
 
 namespace ricaun.Revit.Installation.Tests
@@ -28,6 +29,17 @@ namespace ricaun.Revit.Installation.Tests
             var assemblyFile = Assembly.GetExecutingAssembly().Location;
             Assert.IsFalse(RevitUtils.TryGetRevitVersion(assemblyFile, out int revitVersion));
             Assert.That(revitVersion, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void RevitUtils_Test_GetReferencesRevit()
+        {
+            var fileInfo = new FileInfo("Files/RevitAddin1.dll");
+            var assemblyFile = fileInfo.FullName;
+            Console.WriteLine(assemblyFile);
+            Assert.IsTrue(RevitUtils.TryGetRevitVersion(assemblyFile, out int revitVersion));
+            Assert.That(revitVersion, Is.EqualTo(2017));
+            Console.WriteLine(revitVersion);
         }
     }
 }